+++ src/main/java/org/prebid/server/settings/model/AccountPriceFloorsFetchConfig.java
@@ -19,11 +19,11 @@ public class AccountPriceFloorsFetchConfig {
     Long maxFileSize;
 
     @JsonProperty("max-rules")
-    Integer maxRules;
+    Long maxRules;
 
     @JsonProperty("max-age-sec")
-    Integer maxAgeSec;
+    Long maxAgeSec;
 
     @JsonProperty("period-sec")
-    Integer periodSec;
+    Long periodSec;
 }
